BETA: Fix hero fights history out of range
  * Time need to be long

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/entity/Hero.java
@@ -319,7 +319,7 @@ public class Hero
 				int charTwoClass;
 				int winner;
 				long start;
-				int time;
+				long time;
 				int classed;
 				while (rset.next())
 				{
@@ -329,7 +329,7 @@ public class Hero
 					charTwoClass = rset.getInt("charTwoClass");
 					winner = rset.getInt("winner");
 					start = rset.getLong("start");
-					time = rset.getInt("time");
+					time = rset.getLong("time");
 					classed = rset.getInt("classed");
 					
 					if (charId == charOneId)
